+++ src/main/java/org/library/BookApplication.java
@@ -2,12 +2,12 @@ package org.library;
 
 import org.library.controller.ModeController;
 import org.library.service.BookService;
-import org.library.utils.Executor;
+import org.library.controller.Controller;
 
 public class BookApplication {
     private static final ModeController modeController = new ModeController();
     public static void main(String[] args) {
-        Executor executor = new Executor(new BookService(modeController.selectMode()));
-        while(true) executor.run();
+        Controller controller = new Controller(new BookService(modeController.selectMode()));
+        while(true) controller.run();
     }
 }
